Revert "aaaaaaa"
This reverts commit 93a72c02b7efcf02ad8522527f83350af803c017.

+++ RoboticsProject/Map/Map.h
@@ -25,8 +25,8 @@ public:
 	static int mapWidth;
 	static int mapHeight;
 
-	static float mapResolution;
-	static float gridResolution;
+	static double mapResolution;
+	static double gridResolution;
 	static std::string mapPath;
 
 	std::vector<unsigned char> originalImage;
